chore: change dev dataset locations to EU
dataform cannot currently query tables from datasets in different locations.
Since the dev source dataset is EU the dataform outputs datasets need to be in the EU.
The alternative would be to also set the dev source dataset to US,
but this seems less optimal since it would require us to use different
locations on dev and in staging / production.