chore(reporting): configure dataform to execute daily
In a previous PR (https://github.com/GaloyMoney/lana-bank/pull/1127),
we configured dataform to only execute regulatory report related SQL
when generating reports.  This solves the problem of running into
bigquery limits because it means only the necessary subset of the
pipeline is executed.

Unfortunately, generating the reports was the only way we had
of executing dataform.  This means that only the regulatory report tables
are up to date, other tables like the star-schema tables
we’ve been working on to use in holistics, are stale.

This PR adds two lines to the dataform workflow configuration,
in terraform, to execute the full pipeline once a day.
Since this happens on a global level,
it should not cause any bigquery limits issues.